changes repair predict args, adds cell type arg, adds test for CROTON, moves Cas9Emulation module
