+++ dockers/li_2021_wrapper.py
@@ -37,11 +37,7 @@ def run_croton_prediction(repair_prediction: str, guides: [Cas9]) -> [Cas9]:
     repair_prediction = subprocess.run(command, capture_output=True, text=True, input=encoded)
     # encoding='latin1' is for backwards compatibility.
     results = pickle.loads(codecs.decode(repair_prediction.stdout.encode(), 'base64'), encoding='latin1')
-    #print(results)
-    #exit()
 
-    # TODO currently we loop through key, guide pairs to set results, as we did for the keyed_tuples. We might want to
-    #  save keyed_tuple & guide relationships to be sure the different guides get the correct scores.
     for key, guide in enumerate(guides):
         for t in results:
             if t[0] == key:
