+++ testing/end_to_end/test_data/index.json
@@ -17,7 +17,7 @@
     "results": "4.bed"
   },
     "too many arguments": {
-    "args": "-G semirandom -Target ENST00000461467 -offtargetsTable -non0 -w -T 1 -M NGG --maxMismatches 3 -g 20 -f NN --backbone AGGCTAGTCCGT --replace5P GG -t CODING -n N -R 4 -3 'PRODUCT_SIZE_MIN=150,PRODUCT_SIZE_MAX=290,PRIMER_MIN_SIZE=18,PRIMER_MAX_SIZE=25,PRIMER_OPT_SIZE=22,PRIMER_MIN_TM=57,PRIMER_MAX_TM=63,PRIMER_OPT_TM=60' -A 290 -a 20 --rm1perfOff -r -TDP 50 -TUP 20 -c -SC -taleMin 14 -taleMax 200 -nickaseMin 10 -nickaseMax 31 -offtargetMaxDist 100 -n A -v 0 -p 1 -DF 300 -repairPredictions mESC -filterGCmin 5 -filterGCmax 95 -filterSelfCompMax 1 -consensusUnion -GenBank",
+    "args": "-G semirandom -Target ENST00000461467 -offtargetsTable -non0 -w -T 1 -M NGG --maxMismatches 3 -g 20 -f NN --backbone AGGCTAGTCCGT --replace5P GG -t CODING -n N -R 4 -3 'PRODUCT_SIZE_MIN=150,PRODUCT_SIZE_MAX=290,PRIMER_MIN_SIZE=18,PRIMER_MAX_SIZE=25,PRIMER_OPT_SIZE=22,PRIMER_MIN_TM=57,PRIMER_MAX_TM=63,PRIMER_OPT_TM=60' -A 290 -a 20 --rm1perfOff -r -TDP 50 -TUP 20 -c -SC -taleMin 14 -taleMax 200 -nickaseMin 10 -nickaseMax 31 -offtargetMaxDist 100 -n A -v 0 -p 1 -DF 300 -repairPredictions inDelphi -filterGCmin 5 -filterGCmax 95 -filterSelfCompMax 1 -consensusUnion -GenBank",
     "results": "5.bed"
   },
     "TALEN": {
@@ -37,23 +37,23 @@
     "results": "9.bed"
   },
     "repairPredictions_mESC": {
-    "args": "-Target ENST00000607096 -G semirandom -repairPredictions mESC",
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions inDelphi",
     "results": "10.bed"
   },
     "repairPredictions_U2OS": {
-    "args": "-Target ENST00000607096 -G semirandom -repairPredictions U2OS",
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions inDelphi -cellType U2OS",
     "results": "11.bed"
   },
     "repairPredictions_HEK293": {
-    "args": "-Target ENST00000607096 -G semirandom -repairPredictions HEK293",
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions inDelphi -cellType HEK293",
     "results": "12.bed"
   },
     "repairPredictions_HCT116": {
-    "args": "-Target ENST00000607096 -G semirandom -repairPredictions HCT116",
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions inDelphi -cellType HCT116",
     "results": "13.bed"
   },
     "repairPredictions_K562": {
-    "args": "-Target ENST00000607096 -G semirandom -repairPredictions K562",
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions inDelphi -cellType K562",
     "results": "14.bed"
   },
     "Isoforms": {
@@ -75,5 +75,9 @@
     "scoringMethod_ALL": {
     "args": "-G semirandom -Target ENST00000417324 --scoringMethod ALL",
     "results": "18.bed"
+  },
+    "repairPredictions_CROTON": {
+    "args": "-Target ENST00000607096 -G semirandom -repairPredictions CROTON",
+    "results": "20.bed"
   }
 }
